server/requirements.txt
flask
flask-cors
python-dotenv
requests
spotipy
ollama
gunicorn
cryptography
